Move TooltipComponent
